requirements.txt
aiohttp
asyncio
discord.py
random
requests
pyyaml
